реструктурировал структуру
