+++ gopkg/vpn/wireguard/wireguard_tunnel.go
@@ -6,7 +6,6 @@ import (
 	"golang.zx2c4.com/wireguard/tun/netstack"
 	"net"
 	"net/netip"
-	"syscall"
 )
 
 type WireguardTunnel struct {
@@ -60,16 +59,4 @@ func WireguardTunnelFromConfig(config *WireguardConfig, mtu int) (*WireguardTunn
 
 func (tunnel *WireguardTunnel) CloseTunnel() {
 	tunnel.dev.Close()
-}
-
-func (tunnel *WireguardTunnel) Connect(fd int, sa syscall.Sockaddr) (err error) {
-	return nil
-}
-
-func (tunnel *WireguardTunnel) Read(fd int, p []byte) (n int, err error) {
-	return 0, nil
-}
-
-func (tunnel *WireguardTunnel) Write(fd int, p []byte) (n int, err error) {
-	return 0, nil
 }
